+++ src/gregorian-year-month-day.h
@@ -584,7 +584,7 @@ std::ostringstream&
 ymdh::stream(std::ostringstream& os, r_ssize i) const NOEXCEPT
 {
   ymd::stream(os, i);
-  os << ' ';
+  os << 'T';
   rclock::detail::stream_hour(os, hour_[i]);
   return os;
 }