+++ src/quarterly-year-quarter-day.h
@@ -647,7 +647,7 @@ std::ostringstream&
 yqnqdh<S>::stream(std::ostringstream& os, r_ssize i) const NOEXCEPT
 {
   yqnqd<S>::stream(os, i);
-  os << ' ';
+  os << 'T';
   rclock::detail::stream_hour(os, hour_[i]);
   return os;
 }