+++ tests/testthat/test-gregorian-year-month-weekday.R
@@ -134,6 +134,16 @@ test_that("abbreviated ptype is correct", {
 })
 
 # ------------------------------------------------------------------------------
+# format()
+
+test_that("default formats are correct", {
+  expect_snapshot(format(year_month_weekday(2019)))
+  expect_snapshot(format(year_month_weekday(2019, 1)))
+  expect_snapshot(format(year_month_weekday(2019, 1, 1, 2, 1)))
+  expect_snapshot(format(year_month_weekday(2019, 1, 1, 2, 1, 2, 3, 50, subsecond_precision = "microsecond")))
+})
+
+# ------------------------------------------------------------------------------
 # as.character()
 
 test_that("as.character() works", {